# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "A comprehensive MERN stack API for managing school operations including authentication, attendance, grades, quizzes, and timetables",
    "contact": {
      "name": "API Support",
      "email": "support@schoolmanagement.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/timetable/generate": {
      "post": {
        "tags": ["Timetable"],
        "summary": "Generate timetable",
        "description": "Generate timetable with custom subject allocation (Admin only)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["classSection", "subjectAllocation"],
                "properties": {
                  "classSection": {
                    "type": "string",
                    "example": "10A"
                  },
                  "subjectAllocation": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "faculty": {
                          "type": "string"
                        },
                        "facultyName": {
                          "type": "string"
                        },
                        "subject": {
                          "type": "string"
                        },
                        "periodsPerWeek": {
                          "type": "integer"
                        }
                      }
                    }
                  },
                  "academicYear": {
                    "type": "string",
                    "example": "2024-2025"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Timetable generated successfully"
          },
          "400": {
            "description": "Invalid input or timetable already exists"
          },
          "403": {
            "description": "Access denied"
          }
        }
      }
    },
    "/timetable/my-timetable": {
      "get": {
        "tags": ["Timetable"],
        "summary": "Get student timetable",
        "description": "Get timetable for student's class (Student only)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Timetable retrieved successfully"
          },
          "404": {
            "description": "Timetable not available"
          }
        }
      }
    },
    "/timetable/faculty-schedule": {
      "get": {
        "tags": ["Timetable"],
        "summary": "Get faculty schedule",
        "description": "Get schedule for faculty member (Faculty only)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Schedule retrieved successfully"
          }
        }
      }
    },
    "/timetable/{timetableId}": {
      "delete": {
        "tags": ["Timetable"],
        "summary": "Delete timetable",
        "description": "Delete a timetable (Admin only)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "timetableId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Timetable deleted successfully"
          },
          "404": {
            "description": "Timetable not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}